requirements.txt
Flask==3.0.0
pymongo==4.6.1
openpyxl==3.1.2
pandas==2.2.0
Werkzeug==3.0.1
python-dotenv==1.0.0
